refactor: 함수를 관련성있는 파일 안으로 통합
