+++ src/controller/playRace.js
@@ -1,8 +1,6 @@
 import generateCars from '../model/generateCars.js';
-import showResult from '../view/showResult.js';
+import showResult, { showWinners, showRestart } from '../view/showResult.js';
 import getWinners from '../model/getWinners.js';
-import showWinners from '../view/showWinners.js';
-import showRestart from '../view/showRestart.js';
 
 function sortCars(cars) {
   return cars.sort((a, b) => b.position - a.position);
