--- src/view/showRestart.js
@@ -1,3 +1,0 @@
-export default function showRestart() {
-  document.querySelector('.restart-container').style.display = 'flex';
-}
