--- src/view/showWinners.js
@@ -1,4 +1,0 @@
-export default function showWinners(winners) {
-  const resultContainer = document.querySelector('.game-result-container');
-  resultContainer.insertAdjacentHTML('beforeend', `<div class="winners-name">🏆 최종 우승자: ${[...winners]} 🏆</div>`);
-}
